WIP: Work on comparison w simulink #2

+++ src/control_sys/control_sys.h
@@ -18,9 +18,9 @@
 #define MOTOR_ANG_RATE_THRS 0.05
 
 // values shared by the control system, have to be mutexed
-float filter_current_position;    // taken from the star tracker / gyroscope filter
-float tracking_output_angle;      // taken from the tracking algorithm
-float stabilization_output_angle; // taken from the stabilization algorithm
+double filter_current_position;    // taken from the star tracker / gyroscope filter
+double tracking_output_angle;      // taken from the tracking algorithm
+double stabilization_output_angle; // taken from the stabilization algorithm
 
 /* initialise the control_sys component */
 int init_control_sys(void);
